+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* Add `projectile-before-switch-project-hook`.
 * Add the ability to specify the project type via `.dir-locals.el`.
 * Add support for projects using Midje.
 * Add the ability to create missing tests automatically (controlled via the `projectile-create-missing-test-files` defcustom).